add wrapper to display attribution
Some services like Google Places require to display an attribution. The current version does not support this, as it only has an UL element to render the results to. My version adds a wrapper around the list, and gives options to insert custom markup before or after the list.
I also removed the styling from the script as it belongs to CSS imho.